services/python/feed/requirements.txt
asyncio==3.4.3
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.4.4
idna==2.10
neo-api-client @ git+https://github.com/Kotak-Neo/Kotak-neo-api-v2.git@1533903cb5db1a50892223b404d753dbc8fba50e
numpy==2.1.0
pandas==2.2.3
pyjsparser==2.7.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.3
six==1.16.0
tzdata==2025.3
urllib3==1.26.14
websocket-client==1.8.0
websockets==8.1
